🏗️ Architect: Modularized apts/utils
Refactored the monolithic `apts/utils/__init__.py` utility class into a modular package structure to adhere to the Single Responsibility Principle.

- Split logic into new modules: `coordinates.py`, `date.py`, `graph.py`, `text.py`, and `equipment.py`.
- Consolidated plotting-related utilities into `apts/utils/plot.py` and renamed the class to `PlotUtils`.
- Updated repository-wide internal imports to prefer the new modular functions.
- Maintained a backward-compatible `Utils` shim in `apts/utils/__init__.py` to prevent breaking external APIs.
- Verified stability with the full unit test suite (673 passed).
