+++ apts/opticalequipment/camera/base.py
@@ -10,14 +10,14 @@ class Camera(OutputOpticalEquipment):
 
     @classmethod
     def from_database(cls, entry):
-        from ...utils import Utils, Gender
+        from ...utils import map_conn, map_gender, guess_optical_properties, extract_number, Gender
         brand = entry['brand']
         name = entry['name']
         vendor = f'{brand} {name}'
         ol = entry.get('optical_length', 0)
         mass = entry.get('mass', 0)
-        tt = Utils.map_conn(entry.get('tside_thread'))
-        tg = Utils.map_gender(entry.get('tside_gender'))
+        tt = map_conn(entry.get('tside_thread'))
+        tg = map_gender(entry.get('tside_gender'))
         sw, sh = (entry.get('sensor_width_mm'), entry.get('sensor_height_mm'))
         w, h = (entry.get('width'), entry.get('height'))
